add cert stuff

+++ src/server/ctrlms_ws.cpp
@@ -15,16 +15,17 @@
 #include <rdkx_logger.h>
 #include <safec_lib.h>
 #include <ctrlms_ws.h>
-#include <ctrlm_server_platform.h>
 #include <ctrlm_server_app.h>
 
 #ifdef CTRLMS_WSS_ENABLED
+#include "rdkcertselector.h"
+#include <openssl/ssl.h>
 #include <openssl/pkcs12.h>
-#include <ctrlm_hal_certificate.h>
-#define CTRLMS_WS_CIPHER_LIST       "AES256-SHA256:AES128-GCM-SHA256:AES128-SHA256"
-#define CTRLMS_WS_TLS_CERT_KEY_FILE "/tmp/serverXXXXXX"
-#define CTRLMS_WS_CERT_NAME_LEN     (1024)
-#define CTRLMS_WS_CERT_PW_LEN       (128)
+#define CTRLMS_WS_CIPHER_LIST          "AES256-SHA256:AES128-GCM-SHA256:AES128-SHA256"
+#define CTRLMS_WS_TLS_CERT_KEY_FILE    "/tmp/serverXXXXXX"
+#define CTRLMS_WS_CERT_NAME_LEN        (1024)
+#define CTRLMS_WS_CERT_PW_LEN          (128)
+#define CTRLMS_WS_CERT_FILENAME_PREFIX "file://"
 #endif
 
 typedef enum {
@@ -135,6 +136,9 @@ void *ctrlms_ws_main(void *param) {
    state.app_handle    = NULL;
 
    noPollConnOpts *opts = NULL;
+
+   char tmp_cert[32] = {0};
+   bool cert_valid = false;
    do {
       state.app_handle = ctrlms_ws_load_app(&state, false);
 
@@ -150,51 +154,53 @@ void *ctrlms_ws_main(void *param) {
       }
 
       #ifdef CTRLMS_WSS_ENABLED
-      int cert_key_fd   = -1;
-      FILE *cert_key_fp = NULL;
-      char tmp_cert[32] = {0};
-      int err_store;
+      do {
+         int cert_key_fd   = -1;
+         FILE *cert_key_fp = NULL;
+         int err_store;
 
-      safec_rc = sprintf_s(tmp_cert, sizeof(tmp_cert), "%s", CTRLMS_WS_TLS_CERT_KEY_FILE);
-      if(safec_rc < EOK) {
-         ERR_CHK(safec_rc);
-      }
-
-      umask(0600);
-      cert_key_fd = mkstemp(tmp_cert);
-      if (cert_key_fd == -1)
-      {
-         err_store = errno;
-         XLOGD_ERROR("mkstemp failed: <%s>", strerror(err_store));
-         break;
-      }
-
-      cert_key_fp = fdopen(cert_key_fd, "w");
-      if(cert_key_fp == NULL) {
-         err_store = errno;
-         XLOGD_ERROR("fdopen failed: <%s>", strerror(err_store));
-         if(0 != unlink(&tmp_cert[0])) {
-            err_store = errno;
-            XLOGD_ERROR("failed to remove temp cert <%s>", strerror(err_store));
+         safec_rc = sprintf_s(tmp_cert, sizeof(tmp_cert), "%s", CTRLMS_WS_TLS_CERT_KEY_FILE);
+         if(safec_rc < EOK) {
+            ERR_CHK(safec_rc);
          }
-         break;
-      }
 
-      if(!ctrlms_ws_cert_config(cert_key_fp)) {
-         XLOGD_ERROR("failed to set cert or key, exit");
+         umask(0600);
+         cert_key_fd = mkstemp(tmp_cert);
+         if (cert_key_fd == -1)
+         {
+            err_store = errno;
+            XLOGD_ERROR("mkstemp failed: <%s>", strerror(err_store));
+            break;
+         }
+
+         cert_key_fp = fdopen(cert_key_fd, "w");
+         if(cert_key_fp == NULL) {
+            err_store = errno;
+            XLOGD_ERROR("fdopen failed: <%s>", strerror(err_store));
+            if(0 != unlink(&tmp_cert[0])) {
+               err_store = errno;
+               XLOGD_ERROR("failed to remove temp cert <%s>", strerror(err_store));
+            }
+            break;
+         }
+
+         if(!ctrlms_ws_cert_config(cert_key_fp)) {
+            XLOGD_ERROR("failed to set cert or key");
+            fclose(cert_key_fp);
+            if(0 != unlink(&tmp_cert[0])) {
+               err_store = errno;
+               XLOGD_ERROR("failed to remove temp cert <%s>", strerror(err_store));
+            }
+            break;
+         }
          fclose(cert_key_fp);
-         if(0 != unlink(&tmp_cert[0])) {
-            err_store = errno;
-            XLOGD_ERROR("failed to remove temp cert <%s>", strerror(err_store));
-         }
-         break;
-      }
-      fclose(cert_key_fp);
 
-      // Init OpenSSL
-      SSL_library_init();
-      SSL_load_error_strings();
-      OpenSSL_add_all_algorithms();
+         // Init OpenSSL
+         SSL_library_init();
+         SSL_load_error_strings();
+         OpenSSL_add_all_algorithms();
+         cert_valid = true;
+      } while(0);
       #endif
       
       if(log_enable) {
@@ -206,17 +212,17 @@ void *ctrlms_ws_main(void *param) {
       nopoll_ctx_set_on_ready(g_ctrlms_ws.nopoll_ctx, ctrlms_ws_on_ready, &state);
       nopoll_ctx_set_on_msg(g_ctrlms_ws.nopoll_ctx, ctrlms_ws_on_message, &state);
       
-      #ifdef CTRLMS_WSS_ENABLED
-      nopoll_conn_opts_set_ssl_protocol(opts, NOPOLL_METHOD_TLSV1_2);
-      nopoll_conn_opts_ssl_host_verify(opts, nopoll_false); //localhost will not match host specified in certificate
+      if(cert_valid) {
+         nopoll_conn_opts_set_ssl_protocol(opts, NOPOLL_METHOD_TLSV1_2);
+         nopoll_conn_opts_ssl_host_verify(opts, nopoll_false); //localhost will not match host specified in certificate
 
-      if(!nopoll_conn_opts_set_ssl_certs(opts, &tmp_cert[0], &tmp_cert[0], NULL, NULL)) {
-         XLOGD_ERROR("Failed to add cert/key files to nopoll_conn");
-         nopoll_ctx_unref(g_ctrlms_ws.nopoll_ctx);
-         nopoll_conn_opts_free(opts);
-         break;
+         if(!nopoll_conn_opts_set_ssl_certs(opts, &tmp_cert[0], &tmp_cert[0], NULL, NULL)) {
+            XLOGD_ERROR("Failed to add cert/key files to nopoll_conn");
+            nopoll_ctx_unref(g_ctrlms_ws.nopoll_ctx);
+            nopoll_conn_opts_free(opts);
+            return(NULL);
+         }
       }
-      #endif
 
       char port[6];
       safec_rc = sprintf_s(port, sizeof(port), "%u", port_int);
@@ -225,11 +231,11 @@ void *ctrlms_ws_main(void *param) {
       }
 
       // Start IPv4/6 listener
-      #ifdef CTRLMS_WSS_ENABLED
-      state.nopoll_conn = nopoll_listener_tls_new_opts6(g_ctrlms_ws.nopoll_ctx, opts, "::", port);
-      #else
-      state.nopoll_conn = nopoll_listener_new_opts6(g_ctrlms_ws.nopoll_ctx, opts, "::", port);
-      #endif
+      if(cert_valid) {
+         state.nopoll_conn = nopoll_listener_tls_new_opts6(g_ctrlms_ws.nopoll_ctx, opts, "::", port);
+      } else {
+         state.nopoll_conn = nopoll_listener_new_opts6(g_ctrlms_ws.nopoll_ctx, opts, "::", port);
+      }
       if(!nopoll_conn_is_ok(state.nopoll_conn)) {
          XLOGD_ERROR("Listener connection IPv6 NOT ok");
          nopoll_ctx_unref(g_ctrlms_ws.nopoll_ctx);
@@ -254,12 +260,12 @@ void *ctrlms_ws_main(void *param) {
       nopoll_ctx_unref(g_ctrlms_ws.nopoll_ctx);
       g_ctrlms_ws.nopoll_ctx = NULL;
 
-      #ifdef CTRLMS_WSS_ENABLED
-      if(0 != unlink(tmp_cert)) {
-         int err_store = errno;
-         XLOGD_ERROR("failed to remove temp cert <%s>", strerror(err_store));
+      if(cert_valid) {
+         if(0 != unlink(tmp_cert)) {
+            int err_store = errno;
+            XLOGD_ERROR("failed to remove temp cert <%s>", strerror(err_store));
+         }
       }
-      #endif
    }
 
    if(state.app_handle != NULL) {
@@ -384,18 +390,8 @@ void ctrlms_ws_on_close(noPollCtx *ctx, noPollConn *conn, noPollPtr user_data) {
 #ifdef CTRLMS_WSS_ENABLED
 bool ctrlms_ws_cert_config(FILE* cert_key_fp) {
    bool ret = false;
-
-   ctrlm_hal_certificate_t *hal_certificate = ctrlm_hal_certificate_get();
-
-   if(hal_certificate == NULL) {
-      XLOGD_ERROR("unable to get hal certificate");
-      return(false);
-   }
-
-   ctrlm_voice_cert_t device_cert;
-   bool ocsp_verify_stapling = false;
-   bool ocsp_verify_ca       = false;
-
+   
+   rdkcertselector_h cert_selector = NULL;
    do {
       FILE *device_cert_fp             = NULL;
       PKCS12 *p12_cert                 = NULL;
@@ -403,17 +399,33 @@ bool ctrlms_ws_cert_config(FILE* cert_key_fp) {
       X509 *x509_cert                  = NULL;
       STACK_OF(X509) *additional_certs = NULL;
 
-      if(!hal_certificate->device_cert_get(device_cert, ocsp_verify_stapling, ocsp_verify_ca)) {
-         XLOGD_ERROR("unable to get device certificate");
+      char *cert_path     = NULL;
+      char *cert_password = NULL;
+      cert_selector = rdkcertselector_new(NULL, NULL, "FBK_MTLS");
+
+      if(cert_selector == NULL){
+         XLOGD_TELEMETRY("cert selector init failed");
          break;
       }
 
-      if(device_cert.type != CTRLM_VOICE_CERT_TYPE_P12) {
-         XLOGD_ERROR("unable to parse certificates that are not of PKCS12 type");
+      rdkcertselectorStatus_t cert_status = rdkcertselector_getCert(cert_selector, &cert_path, &cert_password);
+
+      if(cert_status != certselectorOk) {
+         XLOGD_TELEMETRY("cert selector retrieval failed");
+         break;
+      }
+      
+      if(cert_path == NULL || cert_password == NULL) {
+         XLOGD_TELEMETRY("cert selector get failed");
          break;
       }
 
-      device_cert_fp = fopen(device_cert.cert.p12.certificate, "rb");
+      char *local_path = cert_path;
+      if(strncmp(local_path, CTRLMS_WS_CERT_FILENAME_PREFIX, strlen(CTRLMS_WS_CERT_FILENAME_PREFIX)) == 0) {
+         local_path += strlen(CTRLMS_WS_CERT_FILENAME_PREFIX);
+      }
+
+      device_cert_fp = fopen(local_path, "rb");
       if(device_cert_fp == NULL) {
          XLOGD_ERROR("unable to open P12 certificate");
          break;
@@ -428,7 +440,7 @@ bool ctrlms_ws_cert_config(FILE* cert_key_fp) {
          break;
       }
 
-      if(1 != PKCS12_parse(p12_cert, device_cert.cert.p12.passphrase, &pkey, &x509_cert, &additional_certs)) {
+      if(1 != PKCS12_parse(p12_cert, cert_password, &pkey, &x509_cert, &additional_certs)) {
          XLOGD_ERROR("unable to parse P12 certificate");
          break;
       }
@@ -443,15 +455,17 @@ bool ctrlms_ws_cert_config(FILE* cert_key_fp) {
          break;
       }
 
-      if(1 != PEM_write_PrivateKey(cert_key_fp, pkey, NULL, (unsigned char*)device_cert.cert.p12.passphrase, strlen(device_cert.cert.p12.passphrase), NULL, NULL)) {
+      if(1 != PEM_write_PrivateKey(cert_key_fp, pkey, NULL, (unsigned char*)cert_password, strlen(cert_password), NULL, NULL)) {
          XLOGD_ERROR("failed to write temp key");
          break;
-      }
+       }
 
       ret = true;
    }while(0);
 
-   free(hal_certificate);
+   if(cert_selector != NULL) {
+      rdkcertselector_free(&cert_selector);
+   }
 
    return ret;
 }
